make base class more slim
