+++ posthog/hogql_queries/web_analytics/test/test_web_bounces_pre_aggregated.py
@@ -1,5 +1,6 @@
 from freezegun import freeze_time
 
+from posthog.clickhouse.client.execute import sync_execute
 from posthog.models.web_preaggregated.sql import WEB_BOUNCES_INSERT_SQL
 from posthog.models.utils import uuid7
 from posthog.hogql_queries.web_analytics.test.web_preaggregated_test_base import WebAnalyticsPreAggregatedTestBase
@@ -113,8 +114,101 @@ class TestWebBouncesPreAggregated(WebAnalyticsPreAggregatedTestBase):
 
         flush_persons_and_events()
 
+    def _get_pre_agg_metrics_from_bounce_table(self, base_sql: str) -> tuple:
+        metrics_sql = f"""
+        WITH session_data AS (
+            {base_sql}
+        ),
+        aggregated AS (
+            SELECT
+                uniqMerge(persons_uniq_state) AS unique_persons,
+                sumMerge(pageviews_count_state) AS total_pageviews,
+                uniqMerge(sessions_uniq_state) AS unique_sessions,
+                sumMerge(total_session_duration_state) AS total_duration,
+                sumMerge(total_session_count_state) AS total_session_count,
+                sumMerge(bounces_count_state) AS bounce_sessions
+            FROM session_data
+        )
+        SELECT
+            unique_persons,
+            total_pageviews,
+            unique_sessions,
+            if(total_session_count > 0, total_duration / total_session_count, 0) AS avg_session_duration,
+            bounce_sessions,
+            if(unique_sessions > 0, bounce_sessions / unique_sessions, 0) AS bounce_rate
+        FROM aggregated
+        """
+
+        results = sync_execute(metrics_sql)
+        assert len(results) == 1
+        return results[0]
+
+    def _validate_metrics(self, actual_metrics: tuple, expected_metrics: dict):
+        unique_persons, total_pageviews, unique_sessions, avg_duration, bounce_sessions, bounce_rate = actual_metrics
+
+        assert unique_persons == expected_metrics["unique_persons"]
+        assert total_pageviews == expected_metrics["total_pageviews"]
+        assert unique_sessions == expected_metrics["unique_sessions"]
+        assert bounce_sessions == expected_metrics["bounce_sessions"]
+        assert abs(bounce_rate - expected_metrics["bounce_rate"]) < 0.01
+        assert abs(avg_duration - expected_metrics["avg_session_duration"]) < 1.0
+
+    def _get_metrics_from_raw_events(self, date_start: str = "2024-01-01", date_end: str = "2024-01-02") -> tuple:
+        raw_metrics_sql = f"""
+        WITH session_metrics AS (
+            SELECT
+                argMax(if(NOT empty(events__override.distinct_id), events__override.person_id, events.person_id), e.timestamp) AS person_id,
+                countIf(e.event IN ('$pageview', '$screen')) AS pageview_count,
+                e.`$session_id` AS session_id,
+                min(e.timestamp) AS min_timestamp,
+                max(e.timestamp) AS max_timestamp,
+                dateDiff('second', min(e.timestamp), max(e.timestamp)) AS session_duration
+            FROM events AS e
+            LEFT JOIN
+            (
+                SELECT
+                    argMax(person_distinct_id_overrides.person_id, person_distinct_id_overrides.version) AS person_id,
+                    person_distinct_id_overrides.distinct_id AS distinct_id
+                FROM person_distinct_id_overrides
+                WHERE team_id = {self.team.pk}
+                GROUP BY person_distinct_id_overrides.distinct_id
+                HAVING ifNull(argMax(person_distinct_id_overrides.is_deleted, person_distinct_id_overrides.version) = 0, 0)
+            ) AS events__override ON e.distinct_id = events__override.distinct_id
+            WHERE e.team_id = {self.team.pk}
+                AND e.event IN ('$pageview', '$screen')
+                AND e.`$session_id` IS NOT NULL
+                AND e.timestamp >= toDateTime('{date_start}', 'UTC')
+                AND e.timestamp < toDateTime('{date_end}', 'UTC')
+            GROUP BY e.`$session_id`
+        )
+        SELECT
+            uniq(person_id) AS unique_persons,
+            sum(pageview_count) AS total_pageviews,
+            uniq(session_id) AS unique_sessions,
+            if(count(*) > 0, sum(session_duration) / count(*), 0) AS avg_session_duration,
+            sumIf(1, session_duration = 0) AS bounce_sessions,
+            if(count(*) > 0, sumIf(1, session_duration = 0) / count(*), 0) AS bounce_rate
+        FROM session_metrics
+        """
+
+        results = sync_execute(raw_metrics_sql)
+        assert len(results) == 1
+        return results[0]
+
+    def _compare_preagg_vs_raw_metrics(self, preagg_metrics: tuple, raw_metrics: tuple, tolerance: float = 1.0):
+        preagg_persons, preagg_pageviews, preagg_sessions, preagg_duration, preagg_bounces, preagg_bounce_rate = (
+            preagg_metrics
+        )
+        raw_persons, raw_pageviews, raw_sessions, raw_duration, raw_bounces, raw_bounce_rate = raw_metrics
+
+        assert preagg_persons == raw_persons
+        assert preagg_pageviews == raw_pageviews
+        assert preagg_sessions == raw_sessions
+        assert preagg_bounces == raw_bounces
+        assert abs(preagg_duration - raw_duration) < tolerance
+        assert abs(preagg_bounce_rate - raw_bounce_rate) < 0.01
+
     def _get_expected_metrics(self) -> dict:
-        """Return expected metrics for the test scenarios"""
         return {
             "unique_persons": 5,
             "total_pageviews": 11,  # 3+1+4+1+2 = 11
